database/requirements.txt
greenlet==2.0.2
psycopg2-binary==2.9.6
SQLAlchemy==2.0.19
typing_extensions==4.7.1